move openai to common
