+++ Cargo.toml
@@ -23,6 +23,7 @@ members = [
     "src/common/http",
     "src/common/io",
     "src/common/metrics",
+    "src/common/openai",
     "src/common/tracing",
     "src/common/storage",
     "src/common/profile",
