+++ src/common/openai/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "common-openai"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+doctest = false
+test = false
+
+[dependencies]
+# Workspace dependencies
+common-base = { path = "../base" }
+common-exception = { path = "../exception" }
+
+# Github dependencies
+
+# Crates.io dependencies
+openai_api_rust = "0.1.3"
+
+[dev-dependencies]
